requirements.txt
flask>=2.2.5
flask-cors>=4.0.0
Pillow>=9.5.0
requests>=2.31.0
